fix: separate perf group
